@@ -7,23 +7,21 @@
     "@testing-library/react": "^12.0.0",
     "@testing-library/user-event": "^13.2.1",
     "@types/fs-extra": "^9.0.13",
-    "@types/handlebars": "^4.1.0",
     "@types/jest": "^27.0.1",
     "@types/json-schema": "^7.0.11",
     "@types/node": "^16.7.13",
     "@types/node-fetch": "2",
-    "@types/react": "^17.0.20",
-    "@types/react-dom": "^17.0.9",
-    "antd": "^4.17.3",
-    "freemarker.js": "^1.4.0",
+    "@types/react": "^18.0.12",
+    "@types/react-dom": "^18.0.5",
+    "antd": "^4.21.0",
     "fs-extra": "^10.0.0",
     "handlebars": "^4.7.7",
     "json-schema": "^0.4.0",
     "node-fetch": "2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-router-dom": "^6.2.1",
-    "react-scripts": "5.0.0",
+    "react-scripts": "5.0.1",
     "ts-node": "^10.4.0",
     "tslib": "^2.3.1",
     "typescript": "^4.4.2",
