@@ -1,9 +1,9 @@
 // 自动生成的代码块，首行
-// 禁止修改 #HASH<f2ebb63d5d3f3162d6b126b4860bb0ee584413a96876eb3f660703a5993b9b1a>HASH#
+// 禁止修改 #HASH<d51fd4b9e82b5e6e78ddb08058774d658f12860f975738acf2193699d5f71b66>HASH#
 // 如果该代码块被修改，下次执行生成操作时，则不会更新该代码块。
 // 如果修改了代码块，且希望重新生成，可以删除这几行注释，该文件会强制重新生成。
 
-export interface AttachmentEntity {
+export interface SysAttachmentEntity {
   business: string;
   businessId: number;
   createBy: string;
